SXPR-119 Decouple ingestion jobs creation and processing
